+++ src/shared/Updater.cpp
@@ -8,7 +8,8 @@
 #pragma comment(lib, "wininet.lib")
 
 bool CheckForUpdates() {
-    HINTERNET hInternet = InternetOpenA("BetterAngle/4.9.5", INTERNET_OPEN_TYPE_DIRECT, NULL, NULL, 0);
+    g_isCheckingForUpdates = true;
+    HINTERNET hInternet = InternetOpenA("BetterAngle/4.9.6", INTERNET_OPEN_TYPE_DIRECT, NULL, NULL, 0);
     if (!hInternet) return false;
 
     HINTERNET hUrl = InternetOpenUrlA(hInternet, "https://api.github.com/repos/MahanYTT/BetterAngle/releases/latest", NULL, 0, INTERNET_FLAG_RELOAD, 0);
@@ -45,10 +46,19 @@ bool CheckForUpdates() {
             }
             
             g_latestName = L"GitHub Production Release";
+            
+            if (g_latestVersionOnline != "v4.9.6") {
+                g_updateAvailable = true;
+            } else {
+                g_updateAvailable = false;
+            }
+            
+            g_isCheckingForUpdates = false;
             return (g_latestVersion > 4.92f); 
         }
     }
 
+    g_isCheckingForUpdates = false;
     return false;
 }